+++ app/src/main/res/layout/activity_userpage.xml
@@ -19,6 +19,7 @@
             android:backgroundTint="@color/black"
             android:layout_width="80dp"
             android:layout_height="80dp"
+            android:scaleType="fitXY"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="parent"
